Fail late
This changes the bundler to always return manifests for all namespaces and output an error report for them at the end. This does not change the behavior of CI (other than the time it takes to execute) because the bundler will still always return a non-zero error code if any packages are malformed.