Deterministic archive headers
By default, the `tar` crate will copy attributes
such as the modification time (`mtime`) from the
filesystem. By setting the `HeaderMode` to
`Deterministic` on the archive builder, a constant
timestamp is used (stable values are also used for
other files).

The modification time changed was stable on a
local clone of the package repository, but it
would change in CI because GitHub Actions would
do a fresh clone on every run, and Git doesn't
track the modification date.

This should guarantee that the archive we build
are always the same if the actual contents of the
files did not change.

Fixes #2043