Pull READMEs out of full index
Aside from index.json, the bundler creates a second _full index,_ which is an implementation detail of typst.app. This index has grown to 18MB due to the READMEs included within. This makes it annoying to work with:

- Our CDN, Azure Front Door, does not automatically compress files larger than 8MB
- Downloads from JavaScript and WebKit spuriously fail due to keepalives
- The file occupies a relatively large amount of memory once loaded

Instead, we factor out the READMEs into their own folder that can be loaded on demand.

The new full index without the README field is called `index.extra.json`. To prevent disruption of running infrastructure, `index.full.json` is retained in storage but no longer updated. This exception can be removed soon. The new `readmes` folder is also an implementation detail of typst.app and should not be used by third parties.